pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "RuiDevin", email = "rafaelalvaro00@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
]

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"